+++ front/src/modules/settings/components/SettingsNavbar.tsx
@@ -12,6 +12,8 @@ const NavbarSubContainer = styled.div`
   display: flex;
   width: 160px;
   flex-direction: column;
+  margin-top: 41px;
+  margin-left: auto;
 `;
 
 export function SettingsNavbar() {
